TrendingNews/requirements.txt
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cssselect==1.2.0
dnspython==2.3.0
docopt==0.6.2
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.0
flask==2.3.3
idna==3.4
jieba3k==0.35.1
joblib==1.2.0
lxml==4.9.2
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.24.2
pandas==2.0.0
Pillow==9.5.0
pipreqs==0.4.11
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4.1
streamlit==1.30.0
st-annotated-text==4.0.1
tinysegmenter==0.3
tldextract==3.4.0
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.14
yarg==0.1.9
